lib/database.ts: allow OP_RETURN rewind
When the Tx model is saved to the database, the Address model
balance is also updated. the Database.save_tx method will prepare
OP_RETURN vouts for this Address model. Since "OP_RETURN" is
processed/saved as an Address model, it is effectively keeping
track of how many satoshis have been burned by OP_RETURN outputs
on the network.