fix inflationTotal calculation logic, remove sort
Previously was using an extra MongoDB aggregation to gather the
inflationTotal. This was using an extra ~200ms of total CPU time for
generating day, week, and month inflation charts.

This commit consolidates the inflationTotal into the existing MongoDB
aggregation by using `$sum` pipeline operator. Then we use
`Array.reduce()` call to calculate the inflationTotal.

We also remove the sorting by height because it is not needed to
properly render the plots.