add exbitron API request timeout
Recent behavior has been observed where the process that updates
the market database will hang indefinitely. This locks the database
and causes the data to go stale.

It is likely that the issue is with Exibtron API hanging our requests.
This commit adds a default API timeout of 5s to prevent stale db
entries.