Check block hashes vs block heights (#9)
Previous behavior was polling the height from lotusd. This caused the
good block hash to be returned, nullifying the entire check. This commit
fixes that behavior by checking the saved block hash vs the hash
returned from lotusd, therefore confirming whether or not the block
saved to the database had been orphaned.

Co-authored-by: maff1989 <matthew@urgero.net>
Reviewed-on: http://172.16.10.89:3000/maff1989/lotus-explorer/pulls/9